feat: add interface to several packages
