+++ litmus-portal/graphql-server/graph/analytics.resolvers.go
@@ -8,8 +8,6 @@ import (
 
 	"github.com/google/uuid"
 	"github.com/litmuschaos/litmus/litmus-portal/graphql-server/graph/model"
-	analyticsHandler "github.com/litmuschaos/litmus/litmus-portal/graphql-server/pkg/analytics/handler"
-	analyticsOps "github.com/litmuschaos/litmus/litmus-portal/graphql-server/pkg/analytics/ops"
 	"github.com/litmuschaos/litmus/litmus-portal/graphql-server/pkg/authorization"
 	data_store "github.com/litmuschaos/litmus/litmus-portal/graphql-server/pkg/data-store"
 	"github.com/sirupsen/logrus"
@@ -23,7 +21,7 @@ func (r *mutationResolver) CreateDataSource(ctx context.Context, datasource *mod
 		return nil, err
 	}
 
-	return analyticsHandler.CreateDataSource(datasource)
+	return r.analyticsService.CreateDataSource(datasource)
 }
 
 func (r *mutationResolver) CreateDashBoard(ctx context.Context, dashboard *model.CreateDBInput) (*model.ListDashboardResponse, error) {
@@ -34,7 +32,7 @@ func (r *mutationResolver) CreateDashBoard(ctx context.Context, dashboard *model
 		return nil, err
 	}
 
-	return analyticsHandler.CreateDashboard(dashboard)
+	return r.analyticsService.CreateDashboard(dashboard)
 }
 
 func (r *mutationResolver) UpdateDataSource(ctx context.Context, datasource model.DSInput) (*model.DSResponse, error) {
@@ -45,7 +43,7 @@ func (r *mutationResolver) UpdateDataSource(ctx context.Context, datasource mode
 		return nil, err
 	}
 
-	return analyticsHandler.UpdateDataSource(datasource)
+	return r.analyticsService.UpdateDataSource(datasource)
 }
 
 func (r *mutationResolver) UpdateDashboard(ctx context.Context, projectID string, dashboard model.UpdateDBInput, chaosQueryUpdate bool) (string, error) {
@@ -56,11 +54,11 @@ func (r *mutationResolver) UpdateDashboard(ctx context.Context, projectID string
 		return "", err
 	}
 
-	return analyticsHandler.UpdateDashBoard(projectID, dashboard, chaosQueryUpdate)
+	return r.analyticsService.UpdateDashBoard(projectID, dashboard, chaosQueryUpdate)
 }
 
 func (r *mutationResolver) UpdatePanel(ctx context.Context, panelInput []*model.Panel) (string, error) {
-	return analyticsHandler.UpdatePanel(panelInput)
+	return r.analyticsService.UpdatePanel(panelInput)
 }
 
 func (r *mutationResolver) DeleteDashboard(ctx context.Context, projectID string, dbID *string) (bool, error) {
@@ -71,7 +69,7 @@ func (r *mutationResolver) DeleteDashboard(ctx context.Context, projectID string
 		return false, err
 	}
 
-	return analyticsHandler.DeleteDashboard(projectID, dbID)
+	return r.analyticsService.DeleteDashboard(projectID, dbID)
 }
 
 func (r *mutationResolver) DeleteDataSource(ctx context.Context, projectID string, input model.DeleteDSInput) (bool, error) {
@@ -82,7 +80,7 @@ func (r *mutationResolver) DeleteDataSource(ctx context.Context, projectID strin
 		return false, err
 	}
 
-	return analyticsHandler.DeleteDataSource(projectID, input)
+	return r.analyticsService.DeleteDataSource(projectID, input)
 }
 
 func (r *queryResolver) ListHeatmapData(ctx context.Context, projectID string, workflowID string, year int) ([]*model.HeatmapDataResponse, error) {
@@ -93,7 +91,7 @@ func (r *queryResolver) ListHeatmapData(ctx context.Context, projectID string, w
 		return nil, err
 	}
 
-	return analyticsHandler.ListHeatmapData(workflowID, projectID, year)
+	return r.analyticsService.ListHeatmapData(workflowID, projectID, year)
 }
 
 func (r *queryResolver) ListWorkflowStats(ctx context.Context, projectID string, filter model.TimeFrequency, showWorkflowRuns bool) ([]*model.WorkflowStatsResponse, error) {
@@ -104,7 +102,7 @@ func (r *queryResolver) ListWorkflowStats(ctx context.Context, projectID string,
 		return nil, err
 	}
 
-	return analyticsHandler.ListWorkflowStats(projectID, filter, showWorkflowRuns)
+	return r.analyticsService.ListWorkflowStats(projectID, filter, showWorkflowRuns)
 }
 
 func (r *queryResolver) GetWorkflowRunStats(ctx context.Context, workflowRunStatsRequest model.WorkflowRunStatsRequest) (*model.WorkflowRunStatsResponse, error) {
@@ -115,7 +113,7 @@ func (r *queryResolver) GetWorkflowRunStats(ctx context.Context, workflowRunStat
 		return nil, err
 	}
 
-	return analyticsHandler.GetWorkflowRunStats(workflowRunStatsRequest)
+	return r.analyticsService.GetWorkflowRunStats(workflowRunStatsRequest)
 }
 
 func (r *queryResolver) ListDataSource(ctx context.Context, projectID string) ([]*model.DSResponse, error) {
@@ -126,20 +124,20 @@ func (r *queryResolver) ListDataSource(ctx context.Context, projectID string) ([
 		return nil, err
 	}
 
-	return analyticsHandler.QueryListDataSource(projectID)
+	return r.analyticsService.QueryListDataSource(projectID)
 }
 
 func (r *queryResolver) GetPrometheusData(ctx context.Context, request *model.PrometheusDataRequest) (*model.PrometheusDataResponse, error) {
-	promResponseData, _, err := analyticsHandler.GetPrometheusData(request)
+	promResponseData, _, err := r.analyticsService.GetPrometheusData(request)
 	return promResponseData, err
 }
 
 func (r *queryResolver) GetPromLabelNamesAndValues(ctx context.Context, request *model.PromSeriesInput) (*model.PromSeriesResponse, error) {
-	return analyticsHandler.GetLabelNamesAndValues(request)
+	return r.analyticsService.GetLabelNamesAndValues(request)
 }
 
 func (r *queryResolver) GetPromSeriesList(ctx context.Context, request *model.DsDetails) (*model.PromSeriesListResponse, error) {
-	return analyticsHandler.GetPromSeriesList(request)
+	return r.analyticsService.GetPromSeriesList(request)
 }
 
 func (r *queryResolver) ListDashboard(ctx context.Context, projectID string, clusterID *string, dbID *string) ([]*model.ListDashboardResponse, error) {
@@ -150,7 +148,7 @@ func (r *queryResolver) ListDashboard(ctx context.Context, projectID string, clu
 		return nil, err
 	}
 
-	return analyticsHandler.QueryListDashboard(projectID, clusterID, dbID)
+	return r.analyticsService.QueryListDashboard(projectID, clusterID, dbID)
 }
 
 func (r *queryResolver) ListPortalDashboardData(ctx context.Context, projectID string, hubName string) ([]*model.PortalDashboardDataResponse, error) {
@@ -160,7 +158,7 @@ func (r *queryResolver) ListPortalDashboardData(ctx context.Context, projectID s
 	if err != nil {
 		return nil, err
 	}
-	return analyticsHandler.ListPortalDashboardData(projectID, hubName)
+	return r.analyticsService.ListPortalDashboardData(projectID, hubName)
 }
 
 func (r *subscriptionResolver) ViewDashboard(ctx context.Context, dashboardID *string, promQueries []*model.PromQueryInput, dashboardQueryMap []*model.QueryMapForPanelGroup, dataVariables model.DataVars) (<-chan *model.DashboardPromResponse, error) {
@@ -174,13 +172,13 @@ func (r *subscriptionResolver) ViewDashboard(ctx context.Context, dashboardID *s
 		<-ctx.Done()
 		logrus.Printf("Closed dashboard view %v\n", viewID.String())
 		if _, ok := data_store.Store.DashboardData[viewID.String()]; ok {
-			analyticsOps.UpdateViewedAt(dashboardID, viewID.String())
+			r.analyticsService.UpdateViewedAt(dashboardID, viewID.String())
 
 			data_store.Store.Mutex.Lock()
 			delete(data_store.Store.DashboardData, viewID.String())
 			data_store.Store.Mutex.Unlock()
 		}
 	}()
-	go analyticsHandler.DashboardViewer(viewID.String(), dashboardID, promQueries, dashboardQueryMap, dataVariables, *data_store.Store)
+	go r.analyticsService.DashboardViewer(viewID.String(), dashboardID, promQueries, dashboardQueryMap, dataVariables, *data_store.Store)
 	return dashboardData, nil
 }
